Add subsequent exponents scanning to reduce EC point evaluation time
After a random exponent is generated, the corresponding EC point is computed
through complicated arithmetic operations sequence which costs too much.

To reduce the computation there is added ability to use subsequent N exponents
from first generated, in this case the EC point can be easely computed from
previous one by adding the G point (generator) for each exponent increment.

The number of subsequent exponents to be scanned can be defined by means of
`--subseqlen <n>` in the command line. Default is 1,000,000.

When N is 0, the previous (v0.1.0) behaviour is expected
    (this will waste your time!).